# perf(tracer): reduce overhead of attribute sanitization