--- a/packages/opentelemetry-sdk-trace-base/test/performance/benchmark/span.js
+++ b/packages/opentelemetry-sdk-trace-base/test/performance/benchmark/span.js
@@ -18,7 +18,7 @@ const Benchmark = require('benchmark');
 const { BasicTracerProvider } = require('../../../build/src');
 
 const tracerProvider = new BasicTracerProvider();
-const tracer = tracerProvider.getTracer('test')
+const tracer = tracerProvider.getTracer('test');
 
 const suite = new Benchmark.Suite();
 
@@ -26,7 +26,7 @@ suite.on('cycle', event => {
   console.log(String(event.target));
 });
 
-suite.add('create spans (10 attributes)', function() {
+suite.add('create spans (10 attributes)', function () {
   const span = tracer.startSpan('span');
   span.setAttribute('aaaaaaaaaaaaaaaaaaaa', 'aaaaaaaaaaaaaaaaaaaa');
   span.setAttribute('bbbbbbbbbbbbbbbbbbbb', 'aaaaaaaaaaaaaaaaaaaa');
@@ -39,6 +39,23 @@ suite.add('create spans (10 attributes)', function() {
   span.setAttribute('iiiiiiiiiiiiiiiiiiii', 'aaaaaaaaaaaaaaaaaaaa');
   span.setAttribute('jjjjjjjjjjjjjjjjjjjj', 'aaaaaaaaaaaaaaaaaaaa');
   span.end();
+});
+
+suite.add('create spans (10 attributes) passed as options', function () {
+  const span = tracer.startSpan('span', {
+    attributes: {
+      aaaaaaaaaaaaaaaaaaaa: 'aaaaaaaaaaaaaaaaaaaa',
+      bbbbbbbbbbbbbbbbbbbb: 'aaaaaaaaaaaaaaaaaaaa',
+      cccccccccccccccccccc: 'aaaaaaaaaaaaaaaaaaaa',
+      dddddddddddddddddddd: 'aaaaaaaaaaaaaaaaaaaa',
+      eeeeeeeeeeeeeeeeeeee: 'aaaaaaaaaaaaaaaaaaaa',
+      ffffffffffffffffffff: 'aaaaaaaaaaaaaaaaaaaa',
+      gggggggggggggggggggg: 'aaaaaaaaaaaaaaaaaaaa',
+      hhhhhhhhhhhhhhhhhhhh: 'aaaaaaaaaaaaaaaaaaaa',
+      iiiiiiiiiiiiiiiiiiii: 'aaaaaaaaaaaaaaaaaaaa',
+      jjjjjjjjjjjjjjjjjjjj: 'aaaaaaaaaaaaaaaaaaaa',
+    },
+  });
 });
 
 suite.run();