## Which problem is this PR solving?

OTEL is producing too much allocations

Fixes # (issue)

## Short description of the changes

Added possibility for sanitizeAttributes to operate in place on provided object. Remove Object.entries and do prevent copying already allocated copies of arrays

## Type of change

Please delete options that are not relevant.

- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

## How Has This Been Tested?

After:

create spans (10 attributes) x 702,092 ops/sec ±1.32% (88 runs sampled)
create spans (10 attributes) passed as options x 603,744 ops/sec ±0.57% (97 runs sampled)
BatchSpanProcessor process span x 652,096 ops/sec ±0.97% (93 runs sampled)

Before:

create spans (10 attributes) x 691,170 ops/sec ±1.13% (94 runs sampled)
create spans (10 attributes) passed as options x 291,230 ops/sec ±0.54% (99 runs sampled)
BatchSpanProcessor process span x 645,159 ops/sec ±0.97% (95 runs sampled)

Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test configuration

- [ ] Test A

## Checklist:

- [ ] Followed the style guidelines of this project
- [ ] Unit tests have been added
- [ ] Documentation has been updated